# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Produtos e Nutrição",
        "description": "Exemplo Pratico",
        "contact": {
            "name": "Marcos França"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Recupera todos os produtos",
                "description": "Este endpoint recupera uma lista de todos os produtos disponíveis no sistema.",
                "operationId": "b9d77a36a5f7bafc86dc6b2f2c60dd96",
                "responses": {
                    "200": {
                        "description": "Lista de produtos recuperada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "query": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "none"
                                                },
                                                "page": {
                                                    "type": "integer",
                                                    "example": "0"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "product_directory": {
                                            "properties": {
                                                "total_found": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "products": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Erro ao tentar recuperar produtos. Este erro pode ocorrer se houver problemas de conexão com o servidor de banco de dados ou serviço de cache.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "query": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "20221126"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 502
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Erro durante a execução, tente novamente mais tarde."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}